// src/features/app/pages/partner/types/partner-type-data.ts
export type PartnerFormData = {
  month: number;
  monthOfCalculation: string;
  competenceMonth: string;
  contract: number;
  enterprise: string;
  product: string;
  percentageHonorary: number;
  compensation: number;
  honorary: number;
  tax: number;
  tj: number;
  value: number;
  situation: string;
  partnerId: string;
  fee: string;
  feePercentage: string;
  courtFee: string;
};